# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SpeakUp API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Classroom {
  id: number;
  name: string;
  code: string;
  teacherName: string;
  isActive: boolean;
  createdAt: string;
}

export interface ClassroomInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  teacherName: string;
}

export interface ClassSummary {
  classId: number;
  totalQuestions: number;
  answeredQuestions: number;
  unansweredQuestions: number;
}

export interface Question {
  id: number;
  classId: number;
  text: string;
  isAnswered: boolean;
  createdAt: string;
}

export interface QuestionInput {
  /** @minLength 1 */
  text: string;
}

export interface QuestionUpdate {
  isAnswered?: boolean;
}