# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SpeakUp API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: classes
    description: Classroom operations
  - name: questions
    description: Question operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /classes:
    post:
      operationId: createClass
      tags: [classes]
      summary: Create a new classroom
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassroomInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Classroom"
        "400":
          description: Bad request

  /classes/{code}:
    get:
      operationId: getClassByCode
      tags: [classes]
      summary: Get classroom by join code
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Classroom found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Classroom"
        "404":
          description: Not found

  /classes/{classId}/close:
    patch:
      operationId: closeClass
      tags: [classes]
      summary: Close/end a classroom session
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Classroom"
        "404":
          description: Not found

  /classes/{classId}/summary:
    get:
      operationId: getClassSummary
      tags: [classes]
      summary: Get summary stats for a classroom
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassSummary"
        "404":
          description: Not found

  /classes/{classId}/questions:
    get:
      operationId: listQuestions
      tags: [questions]
      summary: List questions for a classroom
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
    post:
      operationId: submitQuestion
      tags: [questions]
      summary: Submit an anonymous question
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "400":
          description: Bad request
        "404":
          description: Classroom not found

  /classes/{classId}/questions/{questionId}:
    patch:
      operationId: updateQuestion
      tags: [questions]
      summary: Mark a question as answered or update it
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: integer
        - name: questionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "404":
          description: Not found

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Classroom:
      type: object
      required: [id, name, code, teacherName, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        teacherName:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string

    ClassroomInput:
      type: object
      required: [name, teacherName]
      properties:
        name:
          type: string
          minLength: 1
        teacherName:
          type: string
          minLength: 1

    ClassSummary:
      type: object
      required: [classId, totalQuestions, answeredQuestions, unansweredQuestions]
      properties:
        classId:
          type: integer
        totalQuestions:
          type: integer
        answeredQuestions:
          type: integer
        unansweredQuestions:
          type: integer

    Question:
      type: object
      required: [id, classId, text, isAnswered, createdAt]
      properties:
        id:
          type: integer
        classId:
          type: integer
        text:
          type: string
        isAnswered:
          type: boolean
        createdAt:
          type: string

    QuestionInput:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1

    QuestionUpdate:
      type: object
      properties:
        isAnswered:
          type: boolean