# lib/api-zod/src/generated/types/classSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SpeakUp API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ClassSummary {
  classId: number;
  totalQuestions: number;
  answeredQuestions: number;
  unansweredQuestions: number;
}