# lib/api-zod/src/generated/types/classroom.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SpeakUp API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Classroom {
  id: number;
  name: string;
  code: string;
  teacherName: string;
  isActive: boolean;
  createdAt: string;
}