# lib/api-zod/src/generated/types/classroomInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SpeakUp API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ClassroomInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  teacherName: string;
}