# lib/api-zod/src/generated/types/question.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SpeakUp API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Question {
  id: number;
  classId: number;
  text: string;
  isAnswered: boolean;
  createdAt: string;
}